Reorganizing code
Version 2 makes the repo cluttered, and difficult to understand/manage.
So I removed it, and added its main changes to the original files.

Amplitude of the thermal diode potential is no longer normalized by the
code in terms of \pi. Any use of \pi should be done by hand.
(as was the case in version "_2", but not on the original one).